+++ src/shared/clients/s3Client.ts
@@ -9,8 +9,6 @@ import { S3 } from 'aws-sdk'
 
 export interface S3Client {
 
-    regionCode: string,
-
     getBucketLocation(bucket: string): Promise<S3.GetBucketLocationOutput>
 
     listBuckets(): Promise<S3.ListBucketsOutput>